enhance: refactor registration form page
